[ADD] add null check before deleting the resource

See merge request choerodon-c7ncd/devops-service!1227

+++ src/main/java/io/choerodon/devops/app/service/impl/DevopsEnvironmentServiceImpl.java
@@ -5,6 +5,7 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 import com.alibaba.fastjson.JSONArray;
+import com.alibaba.fastjson.JSONObject;
 import com.github.pagehelper.PageInfo;
 import com.google.gson.Gson;
 import org.slf4j.Logger;
@@ -965,11 +966,61 @@ public class DevopsEnvironmentServiceImpl implements DevopsEnvironmentService {
     }
 
     @Override
-    public void deletePermissionOfUser(Long envId, Long userId) {
+    public void deletePermissionOfUser(Long projectId, Long envId, Long userId) {
+        if (envId == null || userId == null) {
+            return;
+        }
+
+        UserAttrDTO userAttrDTO = userAttrService.baseQueryById(DetailsHelper.getUserDetails().getUserId());
+
+        if (userAttrDTO == null) {
+            return;
+        }
+
+        DevopsEnvironmentDTO environmentDTO = devopsEnvironmentMapper.selectByPrimaryKey(envId);
+
+        if (environmentDTO == null) {
+            return;
+        }
+
+        if (userAttrDTO.getGitlabUserId() == null) {
+            throw new CommonException("error.gitlab.user.sync.failed");
+        }
+
+        ProjectDTO projectDTO = baseServiceClientOperator.queryIamProjectById(projectId);
+
+        if (baseServiceClientOperator.isProjectOwner(userAttrDTO.getIamUserId(), projectDTO)) {
+            throw new CommonException("error.delete.permission.of.project.owner");
+        }
+
+        // 删除数据库中的纪录
         DevopsEnvUserPermissionDTO devopsEnvUserPermissionDTO = new DevopsEnvUserPermissionDTO();
         devopsEnvUserPermissionDTO.setEnvId(envId);
         devopsEnvUserPermissionDTO.setIamUserId(userId);
         devopsEnvUserPermissionMapper.delete(devopsEnvUserPermissionDTO);
+
+
+        // 删除GitLab对应的项目的权限
+        DevopsEnvUserPayload userPayload = new DevopsEnvUserPayload();
+        userPayload.setIamProjectId(projectId);
+        userPayload.setGitlabProjectId(environmentDTO.getGitlabEnvProjectId().intValue());
+        userPayload.setEnvId(envId);
+        userPayload.setAddGitlabUserIds(Collections.emptyList());
+        // 待删除的用户
+        userPayload.setDeleteGitlabUserIds(Collections.singletonList(userAttrDTO.getGitlabUserId().intValue()));
+        userPayload.setOption(3);
+
+        // 发送saga进行相应的gitlab侧的数据处理
+        producer.apply(
+                StartSagaBuilder.newBuilder()
+                        .withSourceId(projectId)
+                        .withLevel(ResourceLevel.PROJECT)
+                        .withRefType("env")
+                        .withRefId(String.valueOf(envId))
+                        .withJson(JSONObject.toJSONString(userPayload))
+                        .withSagaCode(SagaTopicCodeConstants.DEVOPS_UPDATE_ENV_PERMISSION),
+                builder -> {
+                });
     }
 
     @Override
@@ -1121,10 +1172,15 @@ public class DevopsEnvironmentServiceImpl implements DevopsEnvironmentService {
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public void deleteDeactivatedEnvironment(Long envId) {
+    public void deleteDeactivatedOrFailedEnvironment(Long envId) {
         DevopsEnvironmentDTO devopsEnvironmentDTO = baseQueryById(envId);
-        if (devopsEnvironmentDTO.getActive() != null && devopsEnvironmentDTO.getActive()) {
-            throw new CommonException("error.env.is.active.and.delete.forbidden");
+
+        if (devopsEnvironmentDTO == null) {
+            return;
+        }
+
+        if (!Boolean.FALSE.equals(devopsEnvironmentDTO.getActive()) || !Boolean.TRUE.equals(devopsEnvironmentDTO.getFailed())) {
+            throw new CommonException("error.env.delete");
         }
 
         // 删除环境对应的实例
@@ -1345,7 +1401,7 @@ public class DevopsEnvironmentServiceImpl implements DevopsEnvironmentService {
             devopsEnvAppServiceDTO.setAppServiceId(objectId);
             return devopsEnvAppServiceMapper.selectOne(devopsEnvAppServiceDTO) != null;
         }
-        Boolean check = false;
+        boolean check = false;
         ObjectType objectType = ObjectType.valueOf(type.toUpperCase());
         switch (objectType) {
             case INSTANCE:
